backend/requirements.txt
# HYROXPace Backend Dependencies

# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1

# Redis (for caching, rate limiting)
redis==5.0.1

# HTTP client (for LLM calls)
httpx==0.26.0

# PDF generation (for race plans)
weasyprint==60.2
jinja2==3.1.3

# HTML parsing
beautifulsoup4==4.12.3

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Dev tools
python-dotenv==1.0.0